refactor(desktop): split session lifecycle actions
